Renamer filer og splitter opp komponent
