--- src/component/page/bruker-detaljer/BrukerPaaTiltakDetaljer.module.scss
@@ -1,115 +1,0 @@
-@import '/src/globals.module.scss';
-
-.header {
-	background-color: white;
-	padding: 1.5rem 0;
-
-	border-bottom: 1px solid #c9c9c9
-}
-
-.headerContent {
-	display: grid;
-	grid-template-columns: 1fr;
-	grid-row-gap: 1rem;
-	max-width: $max-bredde;
-	padding: 0 1.5rem;
-	margin: 0 auto;
-
-	@media(min-width: 800px) {
-		grid-template-columns: 100px 1fr;
-		grid-column-gap: 2rem;
-	}
-}
-
-.headerInfoWrapper {
-	display: flex;
-	flex-direction: row;
-	flex-wrap: wrap;
-
-	> *:not(:last-child) {
-		margin-right: 2rem;
-	}
-}
-
-.headerInfo {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-
-	> *:not(:last-child) {
-		margin-right: 2rem;
-	}
-}
-
-.headerTitleWrapper {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-}
-
-.gjennomforingTitle {
-	word-break: break-word;
-}
-
-.headerTitle {
-	margin-right: 2rem;
-}
-
-.tilbakelenkeWrapper {
-	display: flex;
-	justify-content: flex-start;
-	align-items: flex-start;
-}
-
-.tilbakelenke {
-	margin-right: 2rem;
-	margin-top: 0.5rem;
-}
-
-.statusAlert {
-	max-width: 400px;
-}
-
-.skjermetPersonAlert {
-    margin-top: 1rem;
-	max-width: 400px;
-}
-
-.detaljer {
-	display: grid;
-	padding: 2.5rem 1.5rem;
-	max-width: $max-bredde;
-	margin: 0 auto;
-
-	grid-template-columns: 1fr;
-
-	@media(min-width: 800px) {
-		grid-template-columns: 1fr 300px;
-		grid-column-gap: 4rem;
-	}
-}
-
-.deltakerDetaljer {
-	> *:not(:last-child) {
-		margin-bottom: 1.5rem;
-	}
-}
-
-.tiltakCard {
-	width: 100%;
-	margin-bottom: 2rem;
-
-	display: flex;
-	flex-direction: row;
-
-	> :not(:first-child) {
-		margin-left: 4rem;
-	}
-}
-
-.userInfoContent {
-	display: grid;
-	grid-template-columns: 1fr 1fr 1fr;
-	grid-column-gap: 2rem;
-	grid-row-gap: 2rem;
-}
