+++ src/component/page/bruker-detaljer/DeltakerDetaljerHeader.module.scss
@@ -1,0 +1,58 @@
+@import '/src/globals.module.scss';
+
+.header {
+  background-color: white;
+  padding: 1.5rem 0;
+
+  border-bottom: 1px solid #c9c9c9
+}
+
+.headerContent {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-row-gap: 1rem;
+  max-width: $max-bredde;
+  padding: 0 1.5rem;
+  margin: 0 auto;
+
+  @media(min-width: 800px) {
+    grid-template-columns: 100px 1fr;
+    grid-column-gap: 2rem;
+  }
+}
+
+.tilbakelenkeWrapper {
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+}
+
+.headerInfoWrapper {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+
+  > *:not(:last-child) {
+    margin-right: 2rem;
+  }
+}
+
+.headerTitleWrapper {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.headerTitle {
+  margin-right: 2rem;
+}
+
+.headerInfo {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  > *:not(:last-child) {
+    margin-right: 2rem;
+  }
+}
